fix "resource has children" trap due to undropped pollables
Previously, it was possible to drop an `input-stream`, `output-stream`, or
`future-incoming-response` while it still had an undropped child `pollable`
given that `spin-executor` provided no way to cancel a `pollable` previously
registered using `push_waker`.  In addition, it was possible to unintentionally
register more than one pollable for such a resource concurrently.

This commit adds a `CancelToken` return value to `push_waker`, which may be used
to explicitly cancel the registration and immediately drop the `pollable`.  We
now use that, along with `CancelOnDropToken`, to ensure that we cancel any
previous registration before creating a new one, as well as guarantee that the
registration is cancelled before attempting to drop the parent resource.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>