add new `spin_executor::push_waker_and_get_token` function
My previous commit changed the return value of `push_waker`, which was a
breaking change.  Given that it was part of a bugfix that we'd like to make a
point release for, I've converted it to a non-breaking change by switching the
`push_waker` signature back and adding a new function instead.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>